feat: adding anchor and componentizing textual area of ​​sign up
